[TASK] Reformat friwahl-client.pl
This does a pure reformatting without changing any logic in the code,
to make the code better readable.

The reformatting has been performed by perltidy, with "perltidy -b
friwahl-client.pl" and a manual replacement of four spaces with a tab
(because perltidy complained that the option "-t" could not be used).